Reorganize layout component
